refactor: new block model
Now blocks have a new parameter: 'end_pos'. It is the offset form the
start of the line to the end of the block. The unlimited block has
'end_pos' equal to -1.

The function 'grammar_comment#text_block#get_blocks' now returns a
vector with all blocks, no longer a dictionary with 'blocks' and
'unl_blocks'. To check whether the block is unlimited or normal, see
the 'end_pos' parameter of the block.

---
Changes to be committed:
      modified:   autoload/grammar_comment.vim
      modified:   autoload/grammar_comment/functions.vim
      modified:   autoload/grammar_comment/text_block.vim
      modified:   autoload/grammar_comment/vimscript/functions.vim